refactor(core): rename the sign-in-experience-settings class
rename the sign-in-experience-settings class
